+++ modules/Redirect.js
@@ -40,7 +40,7 @@ var Redirect = createClass({
     }
 
   },
-  
+
   propTypes: {
     path: string,
     from: string, // Alias for path
@@ -57,7 +57,7 @@ var Redirect = createClass({
       '<Redirect> elements are for router configuration only and should not be rendered'
     );
   }
-  
+
 });
 
 export default Redirect;
