+++ modules/__tests__/Link-test.js
@@ -2,7 +2,6 @@ import assert from 'assert';
 import expect from 'expect';
 import React from 'react/addons';
 import createHistory from 'history/lib/createMemoryHistory';
-import enableQueries from 'history/lib/enableQueries';
 import execSteps from './execSteps';
 import Router from '../Router';
 import Route from '../Route';
@@ -48,7 +47,7 @@ describe('A <Link>', function () {
     });
 
     React.render((
-      <Router history={enableQueries(createHistory)('/')}>
+      <Router history={createHistory('/')}>
         <Route path="/" component={LinkWrapper}/>
       </Router>
     ), node, function () {
