+++ modules/__tests__/Router-test.js
@@ -1,7 +1,7 @@
 import expect from 'expect';
 import React from 'react';
 import { createLocation } from 'history';
-import Router from '../take-2/Router';
+import Router from '../Router';
 import Route from '../Route';
 
 describe('Router', function () {
