+++ modules/composeMiddleware.js
@@ -1,10 +1,10 @@
-import compose from '../compose';
+import compose from './compose';
 
 export default function composeMiddleware(...middlewares) {
   return routes => {
-    middlewares = middlewares.map(m => m(routes));
+    const finalMiddlewares = middlewares.map(m => m(routes));
     return match => compose(
-      ...middlewares,
+      ...finalMiddlewares,
       match
     );
   };
