+++ modules/createRouter.js
@@ -1,54 +1,90 @@
-import warning from 'warning';
-import matchRoutes from './matchRoutes';
+import identityMiddleware from './identityMiddleware';
+import identityMatch from './identityMatch';
 
-/**
- *  Creates a router.
- *
- *  type RouterState = {
- *    location: Location
- *    routes: Array<Route>,
- *    params: Object
- *  }
- */
-export default function createRouter(initialState) {
-  let state = initialState;
+const defaults = {
+  middleware: identityMiddleware,
+  onError: error => { throw error; }
+};
 
-  /**
-   * Update router state in response to new location
-   * @param {Array<Route>} routes - Route configuration
-   * @param {Location} location - location returned from history
-   * @param {Function} callback - Called with (error, nextState)
-   */
-  function match(routes, location, callback) {
-    matchRoutes(routes, location, (error, nextState) => {
-      if (error) {
-        callback(error);
-        return;
+export default function createRouter(createHistory) {
+  return options => {
+    const {
+      middleware,
+      routes,
+      initialState,
+      onError,
+      ...rest
+    } = { ...defaults, ...options };
+
+    const history = createHistory(rest);
+    let state;
+    let listeners = [];
+
+    const match = middleware(routes)(identityMatch);
+
+    function listen(listener) {
+      listeners.push(listener);
+      listener(state);
+      return function unlisten() {
+        listeners = listeners.filter(l => l !== listener);
+      };
+    }
+
+    function getState() {
+      return state;
+    }
+
+    /**
+     * @private
+     */
+    function setState(nextState) {
+      state = { ...state, ...nextState };
+      listeners.forEach(l => l(state));
+    }
+
+    history.listen(location => {
+      let asyncMatch = true;
+      match({ ...getState(), location }, location, (error, nextState, redirectInfo) => {
+        asyncMatch = false;
+        const { pendingLocation } = getState() || {};
+
+        // Ensure this is the latest location
+        if (pendingLocation && location !== pendingLocation) {
+          return;
+        }
+
+        if (error) {
+          throw onError(error);
+        }
+
+        if (redirectInfo) {
+          history.pushState(
+            redirectInfo.state,
+            redirectInfo.pathname,
+            redirectInfo.query
+          );
+          return;
+        }
+
+        setState({
+          ...nextState,
+          pendingLocation: null
+        });
+      });
+
+      // Only set pending location if match is async
+      if (asyncMatch) {
+        setState({
+          pendingLocation: location
+        });
       }
-      if (!nextState) {
-        warning(
-          false,
-          'Location "%s" did not match any routes',
-          location.pathname + location.search
-        );
-        callback(error, nextState);
-        return;
-      }
-      state = { ...nextState, location };
-      callback(error, state);
     });
-  }
 
-  /**
-   * Get current router state.
-   * @return {RouterState}
-   */
-  function getState() {
-    return state;
-  }
-
-  return {
-    match,
-    getState
+    return {
+      ...history,
+      listen,
+      getState,
+      match
+    };
   };
 }
