+++ modules/identityMiddleware.js
@@ -1,0 +1,6 @@
+/**
+ * A middleware that does nothing
+ */
+export default function identityMiddleware() {
+  return match => match
+}
