+++ modules/routerContext.js
@@ -2,23 +2,18 @@ import React, { Component } from 'react';
 import * as RouterPropTypes from './PropTypes';
 
 /**
- * Higher-order component that adds router and history to context
+ * Higher-order component that adds router to context
  * @param  {Router} router
- * @param  {History} history
  * @return {ReactComponent}
  */
-export default function routerContext(router, history) {
+export default function routerContext(router) {
   return BaseComponent => class RouterContext extends Component {
     static childContextTypes = {
-      router: RouterPropTypes.router,
-      history: RouterPropTypes.history
+      router: RouterPropTypes.router
     }
 
     getChildContext() {
-      return {
-        router,
-        history
-      };
+      return { router };
     }
 
     render() {
