+++ modules/useRoutes.js
@@ -1,23 +1,24 @@
 import warning from 'warning';
-import matchRoutes from '../matchRoutes';
+import matchRoutes from './matchRoutes';
 
 export default function useRoutes(routes) {
   return match => (prevState, location, callback) => {
-    matchRoutes(routes, location, (error1, state) => {
+    matchRoutes(routes, location, (error1, state = { routes: [], params: {} }) => {
       // State from `matchRoutes()` is *not* complete router state
       // Only has `routes` (active routes) and `params`
       if (error1) {
         callback(error1);
         return;
       }
-      if (!state) {
+      if (!routes.length) {
         warning(
           false,
           'Location "%s" did not match any routes',
           location.pathname + location.search
         );
       }
-      match({ prevState, ...state }, location, (error2, nextState, redirectInfo) => {
+
+      match({ ...prevState, ...state }, location, (error2, nextState, redirectInfo) => {
         if (error2 || redirectInfo) {
           callback(error2, null, redirectInfo);
           return;
